# Optimize INTERPOLATE even more