Make as many parameters anonymous or non-sigiled as possible.

For a 10k line file, `./perl6-m -e 'my @l = "sm.sql".IO.lines; my $s = "Perl6"; my $t = now; my @m = @l.grep(/ $s /); say @m.elems; say now - $t'` used to take ~2.7s, now it takes ~2.2s. `/ <$s> /` used to take ~11.4s, now it takes ~8.4s. `my @s = <Perl6 Perl6 Perl6 Perl6>; my $t = now; my @m = @l.grep(/ @s /);` used to take ~159s, now it takes ~110s.

Rakudo builds ok and passes make m-test m-spectest.