@AlexDaniel 2017-10-21
That's a bit close to the release but I'll take it. :) Thanks!